TAJO-781: Implement expression transformation for merge phase
Reviewed-on: https://dbserver.korea.ac.kr/reviews/222
Tested-by:  Jenkins <build@dbserver.korea.ac.kr>

+++ engine/src/main/java/nta/engine/function/builtin/NewSumInt.java
@@ -16,7 +16,7 @@ public class NewSumInt extends AggFunction<Datum> {
 
   public NewSumInt() {
     super(new Column[] {
-        new Column("val", CatalogProtos.DataType.LONG)
+        new Column("val", CatalogProtos.DataType.INT)
     });
   }
